Add a new base class for structured parsers.
JSON and XML parsers cannot parse flat text files.
