+++ src/test/java/edu/hm/hafner/analysis/parser/DScannerParserTest.java
@@ -2,8 +2,6 @@ package edu.hm.hafner.analysis.parser;
 
 import org.junit.jupiter.api.Test;
 
-import edu.hm.hafner.analysis.AbstractParserTest;
-import edu.hm.hafner.analysis.ParsingException;
 import edu.hm.hafner.analysis.Report;
 import edu.hm.hafner.analysis.Severity;
 import edu.hm.hafner.analysis.assertions.SoftAssertions;
@@ -15,7 +13,7 @@ import static edu.hm.hafner.analysis.assertions.Assertions.*;
  *
  * @author Andre Pany
  */
-class DScannerParserTest extends AbstractParserTest {
+class DScannerParserTest extends StructuredFileParserTest {
     DScannerParserTest() {
         super("dscanner-report.json");
     }
@@ -59,23 +57,10 @@ class DScannerParserTest extends AbstractParserTest {
     }
 
     @Test
-    void shouldNotAcceptTextFiles() {
-        assertThat(createParser().accepts(createReaderFactory("gcc.txt"))).isFalse();
-    }
-
-    @Test
     void shouldHandleIncompleteReports() {
         Report report = createParser().parse(createReaderFactory("dscanner-incomplete-report.json"));
         assertThat(report).hasSize(2);
         assertThat(report.getErrorMessages()).isEmpty();
-    }
-
-    @Test
-    void shouldThrowParserException() {
-        assertThatThrownBy(() -> createParser().parse(createReaderFactory("issues-invalid.json")))
-                .isInstanceOf(ParsingException.class);
-        assertThatThrownBy(() -> createParser().parse(createReaderFactory("issues-broken.json")))
-                .isInstanceOf(ParsingException.class);
     }
 
     @Override
