+++ src/test/java/edu/hm/hafner/analysis/parser/EclipseXMLParserTest.java
@@ -2,7 +2,6 @@ package edu.hm.hafner.analysis.parser;
 
 import org.junit.jupiter.api.Test;
 
-import edu.hm.hafner.analysis.AbstractParserTest;
 import edu.hm.hafner.analysis.Categories;
 import edu.hm.hafner.analysis.Report;
 import edu.hm.hafner.analysis.Severity;
@@ -15,7 +14,7 @@ import static edu.hm.hafner.analysis.assertions.Assertions.*;
  *
  * @author Jason Faust
  */
-class EclipseXMLParserTest extends AbstractParserTest {
+class EclipseXMLParserTest extends StructuredFileParserTest {
     EclipseXMLParserTest() {
         super("eclipse-withinfo.xml");
     }
