+++ src/test/java/edu/hm/hafner/analysis/parser/XmlParserTest.java
@@ -4,7 +4,6 @@ import java.util.Iterator;
 
 import org.junit.jupiter.api.Test;
 
-import edu.hm.hafner.analysis.AbstractParserTest;
 import edu.hm.hafner.analysis.Issue;
 import edu.hm.hafner.analysis.LineRange;
 import edu.hm.hafner.analysis.ParsingException;
@@ -19,7 +18,7 @@ import static org.assertj.core.api.Assertions.*;
  *
  * @author Raphael Furch
  */
-class XmlParserTest extends AbstractParserTest {
+class XmlParserTest extends StructuredFileParserTest {
     private static final String ISSUES_DEFAULT_FILE = "xmlParserDefault.xml";
     private static final String ISSUES_CUSTOM_PATH_FILE = "xmlParserCustomPath.xml";
     private static final String ISSUES_INCOMPATIBLE_VALUE = "xmlParserIncompatibleValues.xml";
@@ -117,8 +116,7 @@ class XmlParserTest extends AbstractParserTest {
     }
 
     @Test
-    void shouldNotAcceptTextFiles() {
-        assertThat(createParser().accepts(createReaderFactory("gcc.txt"))).isFalse();
+    void shouldAcceptSampleFile() {
         assertThat(createParser().accepts(createReaderFactory(ISSUES_DEFAULT_FILE))).isTrue();
     }
